Code refactoring.
Improve inefficient DB queries.
Generalize mixins.
Remove useless modules.
