+++ bridge_adaptivity/module/policies/policy_trials_count.py
@@ -6,4 +6,4 @@ class TrialsCountGradingPolicy(BaseGradingPolicy):
 
     def _calculate(self):
         trials_count, _ = self._get_points_earned_trials_count()
-        return self._internal_calculate(trials_count)
+        return round(float(trials_count) / max(self.policy.threshold, trials_count), 2)
